+++ packages/router-plugin/internals/src/symbols.ts
@@ -1,5 +1,9 @@
 import { InjectionToken } from '@angular/core';
 
+declare const ngDevMode: boolean;
+
+const NG_DEV_MODE = typeof ngDevMode === 'undefined' || ngDevMode;
+
 export const enum NavigationActionTiming {
   PreActivation = 1,
   PostActivation = 2
@@ -9,17 +13,17 @@ export interface NgxsRouterPluginOptions {
   navigationActionTiming?: NavigationActionTiming;
 }
 
-export const USER_OPTIONS = new InjectionToken<NgxsRouterPluginOptions | undefined>(
-  'USER_OPTIONS',
+export const ɵUSER_OPTIONS = new InjectionToken<NgxsRouterPluginOptions | undefined>(
+  NG_DEV_MODE ? 'USER_OPTIONS' : '',
   { providedIn: 'root', factory: () => undefined }
 );
 
-export const NGXS_ROUTER_PLUGIN_OPTIONS = new InjectionToken<NgxsRouterPluginOptions>(
-  'NGXS_ROUTER_PLUGIN_OPTIONS',
+export const ɵNGXS_ROUTER_PLUGIN_OPTIONS = new InjectionToken<NgxsRouterPluginOptions>(
+  NG_DEV_MODE ? 'NGXS_ROUTER_PLUGIN_OPTIONS' : '',
   { providedIn: 'root', factory: () => ({}) }
 );
 
-export function createRouterPluginOptions(
+export function ɵcreateRouterPluginOptions(
   options: NgxsRouterPluginOptions | undefined
 ): NgxsRouterPluginOptions {
   return {