+++ packages/router-plugin/src/router.module.ts
@@ -1,14 +1,14 @@
 import { ModuleWithProviders, NgModule } from '@angular/core';
 import { NgxsModule } from '@ngxs/store';
+import {
+  NgxsRouterPluginOptions,
+  ɵcreateRouterPluginOptions,
+  ɵNGXS_ROUTER_PLUGIN_OPTIONS,
+  ɵUSER_OPTIONS
+} from '@ngxs/router-plugin/internals';
 
 import { RouterState } from './router.state';
 import { DefaultRouterStateSerializer, RouterStateSerializer } from './serializer';
-import {
-  USER_OPTIONS,
-  NGXS_ROUTER_PLUGIN_OPTIONS,
-  NgxsRouterPluginOptions,
-  createRouterPluginOptions
-} from './symbols';
 
 @NgModule({
   imports: [NgxsModule.forFeature([RouterState])]
@@ -20,11 +20,11 @@ export class NgxsRouterPluginModule {
     return {
       ngModule: NgxsRouterPluginModule,
       providers: [
-        { provide: USER_OPTIONS, useValue: options },
+        { provide: ɵUSER_OPTIONS, useValue: options },
         {
-          provide: NGXS_ROUTER_PLUGIN_OPTIONS,
-          useFactory: createRouterPluginOptions,
-          deps: [USER_OPTIONS]
+          provide: ɵNGXS_ROUTER_PLUGIN_OPTIONS,
+          useFactory: ɵcreateRouterPluginOptions,
+          deps: [ɵUSER_OPTIONS]
         },
         { provide: RouterStateSerializer, useClass: DefaultRouterStateSerializer }
       ]